Crate utoipa_gen

This is private utoipa codegen library and is not used alone.

The library contains macro implementations for utoipa library. Content of the library documentation is available through utoipa library itself. Consider browsing via the utoipa crate so all links will work correctly.

Macros

schema

Create OpenAPI Schema from arbitrary type.

Attribute Macros

path

Path attribute macro implements OpenAPI path for the decorated function.

Derive Macros

IntoParams

Generate path parameters from struct’s fields.

IntoResponses

Generate responses with status codes what can be attached to the utoipa::path.

OpenApi

Generate OpenApi base object with defaults from project settings.

ToResponse

Generate reusable OpenAPI response that can be used in utoipa::path or in OpenApi.

ToSchema

Generate reusable OpenAPI schema to be used together with OpenApi.